refactor: consolidate chunking module structure
Major restructuring to improve code organization:

1. Created unified chunking/ directory structure:
   - chunking.rs → chunking/base.rs
   - overlap_chunking/ → chunking/overlap/

2. Consolidated overlap module files (7 → 3 files):
   - manager.rs: merged overlap_chunk_manager.rs + state_tracker.rs
   - processor.rs: merged overlap_processor.rs + pattern_detector.rs + constants.rs
   - types.rs: kept as-is for all type definitions

3. Benefits:
   - Clearer hierarchy between base and overlap chunking
   - Reduced file count improves maintainability
   - Better cohesion of related functionality
   - Preserved backward compatibility with re-exports

All tests pass and CI checks are green.
