+++ sakurs-core/src/application/chunking/base.rs
