+++ sakurs-core/src/application/chunking/overlap/mod.rs
@@ -3,16 +3,13 @@
 //! This module provides an overlap-based chunking system that can detect and handle
 //! patterns (like contractions and possessives) that span chunk boundaries.
 
-mod constants;
-mod overlap_chunk_manager;
-mod overlap_processor;
-mod pattern_detector;
-mod state_tracker;
+mod manager;
+mod processor;
 mod types;
 
-pub use overlap_chunk_manager::{OverlapChunkConfig, OverlapChunkManager};
-pub use overlap_processor::OverlapProcessor;
-pub use state_tracker::{CrossChunkStateTracker, StateTrackerConfig};
+// Re-export public types
+pub use manager::{OverlapChunkConfig, OverlapChunkManager, StateTrackerConfig};
+pub use processor::OverlapProcessor;
 pub use types::{
     BoundaryAdjustment, ChunkTransitionState, OverlapResult, PartialPattern, PatternType,
     ProcessedChunk, SuppressionMarker, SuppressionReason,
