+++ sakurs-core/src/application/chunking/overlap/types.rs
